removed questionaire table. updated storybook to directly use the generic table definition
