+++ components/styled/table/src/index.tsx
@@ -1,20 +1,27 @@
 import { FC } from 'react'
+import GenericTable, { TableData } from './atoms/table'
 
-import { Questionnaire, QuestionnaireResponse, SummaryTableViewType } from '@ltht-react/types'
-import QuestionnaireTable from './molecules/questionnaire-table'
-
-const Table: FC<IProps> = ({ definition, records, orientation = 'VERTICAL' }) => {
-  if (!definition.item || definition.item.length === 0) {
+const Table: FC<IProps> = ({ columnData, rowData, mapToTableData }) => {
+  if (
+    !columnData ||
+    !rowData ||
+    (Array.isArray(columnData) && columnData?.length === 0) ||
+    (Array.isArray(rowData) && rowData?.length === 0)
+  ) {
     return <div>Could not render table. Definition items array was empty.</div>
   }
 
-  return <QuestionnaireTable definitionItems={definition.item} records={records} orientation={orientation} />
+  if (typeof mapToTableData !== 'function') {
+    return <div>Could not render table. mapping method not provided.</div>
+  }
+
+  return <GenericTable columnData={columnData} rowData={rowData} mapToTableData={mapToTableData} />
 }
 
-interface IProps {
-  orientation?: SummaryTableViewType
-  definition: Questionnaire
-  records: QuestionnaireResponse[]
+interface IProps<TColumn = any, TRow = any> {
+  columnData: TColumn
+  rowData: TRow
+  mapToTableData: (colItems: TColumn, rowItems: TRow) => TableData
 }
 
 export default Table
